scripts/requirements.txt
requests==2.31.0
pandas==2.1.0
pyarrow==14.0.1
boto3==1.34.0
selenium==4.18.1
webdriver_manager==4.0.1
python-dotenv==1.0.1 